requirements.txt
psycopg2_binary==2.9.10
streamlit==1.43.2
supabase==2.13.0
